Refactor controls into clock
