+++ src/lib/components/Clock.svelte
@@ -1,13 +1,98 @@
 <script>
 	import { formatDuration } from '$lib/utils/time';
-	import { duration } from '$lib/stores';
+	import { times, duration } from '$lib/stores';
+
+	let modalOpen = $state(false);
+	let category = $state('');
+
+	/**
+	 * @type {import("$lib/stores").Task | undefined}
+	 */
+	let activeTask;
+
+	/**
+	 * Toggle the modal.
+	 */
+	function toggleModal() {
+		modalOpen = !modalOpen;
+	}
+
+	/**
+	 * Function to log the stopwatch
+	 */
+	async function updateTask() {
+		activeTask = $times.tasks?.find((t) => t.status === 'ACTIVE');
+		if (!activeTask) {
+			alert('Please try again.');
+			console.error('Active task expected but not found');
+			return;
+		}
+		try {
+			const res = await fetch(`/api/tasks/${activeTask.id}`, {
+				method: 'PUT',
+				body: JSON.stringify({ category })
+			});
+			if (res.ok) {
+				const { updatedTask, newTask } = await res.json();
+				updateTaskStore(updatedTask, newTask);
+			} else {
+				alert('Failed to add task');
+			}
+		} catch (error) {
+			console.error(error);
+		}
+
+		category = ''; // Disable pre-population next tune
+		toggleModal(); // Close the time log modal
+	}
+
+	/**
+	 * @param {import("$lib/stores").Task} updatedTask
+	 * @param {import("$lib/stores").Task} newTask
+	 */
+	function updateTaskStore(updatedTask, newTask) {
+		times.update((current) => {
+			let newTasks = current.tasks.map((task) => {
+				if (task.id === updatedTask.id) {
+					return {
+						...task,
+						...updatedTask // Merge current with new properties
+					};
+				}
+				// Return unchanged task if ID's don't match
+				return task;
+			});
+			newTasks.push(newTask);
+			return {
+				tasks: newTasks,
+				error: current.error
+			};
+		});
+	}
 </script>
 
 <div class="stopwatch">
-	<p>
-		{formatDuration($duration)}
-	</p>
+	<button onclick={toggleModal} class="submit">{formatDuration($duration)}</button>
 </div>
+
+{#if modalOpen}
+	<div class="modal-background" onclick={toggleModal}>
+		<div class="modal-content" onclick={e => e.stopPropagation()}>
+			<h2>Track Time</h2>
+			<p>You are about to track {formatDuration($duration)}.</p>
+
+			<form onsubmit={e => {e.preventDefault(); updateTask()}}>
+				<div class="category">
+					<label>
+						Task Category:
+						<input type="text" bind:value={category} required />
+					</label>
+				</div>
+				<button class="add-task" type="submit">Add Task</button>
+			</form>
+		</div>
+	</div>
+{/if}
 
 <style>
 	.stopwatch {
@@ -16,8 +101,75 @@
 		display: flex;
 		justify-content: center; /* Centers content horizontally */
 		align-items: center; /* Centers content vertically */
-		border: 1px solid #000; /* Optional: for visualization */
 		box-sizing: content-box;
 		border-radius: 5px;
+
+		font-size: 2em;
+		padding-bottom: 10px;
+		text-align: center;
+	}
+
+	/* Buttons */
+	button {
+		/* padding: 10px 20px; */
+		/* font-size: 1em; */
+		cursor: pointer;
+		width: 100%;
+		height: 100%;
+		border: 1px solid rgb(42, 185, 66);
+		border-radius: 1rem;
+		box-shadow: 1px 1px 1px rgb(70, 194, 91);
+		transform: translate3d(0,0,0); /* Hack to fix poor box-shadow rendering on iOS*/
+	}
+
+	/* Submit button */
+	.submit {
+		background-color: rgb(53, 226, 79);
+		color: white;
+		width: 100%;
+		height: 100%;
+	}
+
+	/* Modal background */
+	.modal-background {
+		z-index: 999;
+		position: fixed;
+		top: 0;
+		left: 0;
+		width: 100%;
+		height: 100%;
+		background-color: rgba(0, 0, 0, 0.5);
+		display: flex;
+		justify-content: center;
+		align-items: center;
+	}
+
+	/* Modal content */
+	.modal-content {
+		background-color: white;
+		padding: 20px;
+		border-radius: 5px;
+		max-width: 500px;
+		width: 100%;
+	}
+	/* Close button */
+	.add-task {
+		cursor: pointer;
+		background-color: #0ccf61;
+		color: white;
+		border: none;
+		padding: 5px 10px;
+		margin-left: auto;
+		border-radius: 3px;
+		min-height: 3rem;
+		font-size: 1.5rem;
+	}
+
+	.category {
+		padding-bottom: 20px;
+	}
+
+	.category input {
+		min-height: 50px;
 	}
 </style>
